services/payment-webhook/requirements.txt
# FastAPI and web framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.7

# HTTP client for API calls
httpx==0.26.0
aiohttp==3.9.3

# Payment providers
stripe==8.2.0
# Strike doesn't have an official Python SDK, we'll use httpx

# Database
aiosqlite==0.19.0

# Email
sendgrid==6.11.0
python-decouple==3.8

# PGP encryption
python-gnupg==0.5.2

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.1

# Logging and monitoring
structlog==24.1.0

# Testing (optional, for development)
pytest==8.0.0
pytest-asyncio==0.23.5
